Fixed CourseInfoModal issues and reorganized contexts
