+++ src/app/components/majorSelection/miniPlanner/MiniQuarterCard.tsx
@@ -1,5 +1,5 @@
 import { isCustomCourse } from "@/lib/plannerUtils";
-import { ModalsContext } from "@contexts/ModalsProvider";
+import { CourseInfoContext } from "@contexts/CourseInfoProvider";
 import { CourseTerm, StoredCourse } from "@customTypes/Course";
 import { Quarter } from "@customTypes/Quarter";
 import { Card, Link, Typography } from "@mui/joy";
@@ -33,7 +33,8 @@ function MiniCourseCard({
   course: StoredCourse;
   quarter: Quarter;
 }) {
-  const { onShowCourseInfoModal, setDisplayCourse } = useContext(ModalsContext);
+  const { onShowCourseInfoModal, setDisplayCourse } =
+    useContext(CourseInfoContext);
 
   function courseTitle(course: StoredCourse) {
     if (course.departmentCode && course.number) {
