+++ src/app/components/planner/Planner.tsx
@@ -1,7 +1,8 @@
-import { MajorVerificationContext } from "@/app/contexts/MajorVerificationProvider";
-import { PermissionsProvider } from "@/app/contexts/PermissionsProvider";
 import { findCoursesInQuarter, quartersPerYear } from "@/lib/plannerUtils";
+import { CourseInfoProvider } from "@contexts/CourseInfoProvider";
+import { MajorVerificationContext } from "@contexts/MajorVerificationProvider";
 import { ModalsContext, ModalsProvider } from "@contexts/ModalsProvider";
+import { PermissionsProvider } from "@contexts/PermissionsProvider";
 import { PlannerContext } from "@contexts/PlannerProvider";
 import { PlannerData } from "@customTypes/Planner";
 import { Quarter } from "@customTypes/Quarter";
@@ -44,49 +45,51 @@ export default function Planner({ isActive }: { isActive: boolean }) {
       <DragDropContext onDragEnd={handleDragEnd}>
         <PermissionsProvider>
           <ModalsProvider>
-            <div className="flex justify-between space-x-4">
-              <div className="flex-initial pr-2">
-                <SearchContainer />
-              </div>
-              <div className="overflow-auto w-full flex-grow">
-                <AccordionGroup>
-                  <div className="space-y-2 h-[75vh] overflow-auto">
-                    <Years courseState={courseState} />
-                    <Accordion
-                      variant="soft"
-                      sx={{
-                        borderRadius: "0.5rem",
-                        "&.MuiAccordion-root": {
-                          "& .MuiAccordionSummary-root": {
-                            padding: "0.5rem 0",
-                            paddingX: "0.5rem",
+            <CourseInfoProvider>
+              <div className="flex justify-between space-x-4">
+                <div className="flex-initial pr-2">
+                  <SearchContainer />
+                </div>
+                <div className="overflow-auto w-full flex-grow">
+                  <AccordionGroup>
+                    <div className="space-y-2 h-[75vh] overflow-auto">
+                      <Years courseState={courseState} />
+                      <Accordion
+                        variant="soft"
+                        sx={{
+                          borderRadius: "0.5rem",
+                          "&.MuiAccordion-root": {
+                            "& .MuiAccordionSummary-root": {
+                              padding: "0.5rem 0",
+                              paddingX: "0.5rem",
+                            },
                           },
-                        },
-                      }}
-                      defaultExpanded={true}
-                    >
-                      <AccordionSummary>Notes</AccordionSummary>
-                      <AccordionDetails>
-                        <NotesEditor
-                          content={courseState.notes}
-                          onUpdateNotes={updateNotes}
-                        />
-                      </AccordionDetails>
-                    </Accordion>
-                  </div>
-                </AccordionGroup>
-              </div>
+                        }}
+                        defaultExpanded={true}
+                      >
+                        <AccordionSummary>Notes</AccordionSummary>
+                        <AccordionDetails>
+                          <NotesEditor
+                            content={courseState.notes}
+                            onUpdateNotes={updateNotes}
+                          />
+                        </AccordionDetails>
+                      </Accordion>
+                    </div>
+                  </AccordionGroup>
+                </div>
 
-              <div className="flex flex-col self-start gap-3">
-                <PlannerActions />
-                <GraduationProgressCard
-                  totalCredits={totalCredits}
-                  geSatisfied={geSatisfied}
-                  courseState={courseState}
-                />
+                <div className="flex flex-col self-start gap-3">
+                  <PlannerActions />
+                  <GraduationProgressCard
+                    totalCredits={totalCredits}
+                    geSatisfied={geSatisfied}
+                    courseState={courseState}
+                  />
+                </div>
               </div>
-            </div>
-            <Modals />
+              <Modals />
+            </CourseInfoProvider>
           </ModalsProvider>
         </PermissionsProvider>
       </DragDropContext>
