+++ src/app/components/planner/graduationProgress/majorProgressModal/FulfillmentMark.tsx
@@ -1,7 +1,7 @@
-import IsSatisfiedMark from "@/app/components/miscellaneous/IsSatisfiedMark";
-import { MajorVerificationContext } from "@/app/contexts/MajorVerificationProvider";
-import { PlannerContext } from "@/app/contexts/PlannerProvider";
-import { Requirements } from "@/app/types/Requirements";
+import IsSatisfiedMark from "@components/miscellaneous/IsSatisfiedMark";
+import { MajorVerificationContext } from "@contexts/MajorVerificationProvider";
+import { PlannerContext } from "@contexts/PlannerProvider";
+import { Requirements } from "@customTypes/Requirements";
 import { useContext } from "react";
 
 export default function FulfillmentMark(requirements: Requirements) {
