+++ src/app/contexts/MajorVerificationProvider/index.ts
@@ -1,0 +1,4 @@
+export {
+  MajorVerificationProvider,
+  MajorVerificationContext,
+} from "./Provider";
