+++ src/app/contexts/ModalsProvider/index.ts
@@ -1,0 +1,1 @@
+export { ModalsProvider, ModalsContext } from "./Provider";
