+++ src/app/contexts/NextAuthProvider.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { SessionProvider } from "next-auth/react";
 
 export default function NextAuthProvider({
