+++ src/app/contexts/PermissionsProvider/Types.ts
@@ -1,0 +1,11 @@
+import { Permissions } from "@customTypes/Permissions";
+
+export interface PermissionsContextProps {
+  loadingPermissions: boolean;
+  isSaved: boolean;
+  permissionsList: Permissions[];
+  onSetPermissionsList: (permissions: Permissions[]) => void;
+  onSavePermissions: () => void;
+  isAdmin: boolean;
+  hasPermissionToEdit: boolean;
+}
