+++ src/app/contexts/PermissionsProvider/index.ts
@@ -1,0 +1,1 @@
+export { PermissionsProvider, PermissionsContext } from "./Provider";
