+++ src/app/contexts/PermissionsProvider/usePermissions.ts
@@ -1,10 +1,9 @@
-import useUserPermissions from "@/app/hooks/useUserPermissions";
-import { Permissions } from "@/app/types/Permissions";
+import { getPermissions, savePermissions } from "@/app/actions/permissions";
+import { Permissions } from "@customTypes/Permissions";
+import useUserPermissions from "@hooks/useUserPermissions";
 import { useMutation, useQuery } from "@tanstack/react-query";
 import { useSession } from "next-auth/react";
 import { useState } from "react";
-
-import { getPermissions, savePermissions } from "../../actions/permissions";
 
 export default function usePermissions() {
   const [permissionsList, setPermissionsList] = useState<Permissions[]>([]);
