+++ src/app/types/Common.ts
@@ -1,1 +1,2 @@
 export type SetState<T> = React.Dispatch<React.SetStateAction<T>>;
+export type SetShow = SetState<boolean>;
